+++ src/index_profile/src/calls.rs
@@ -1,10 +1,11 @@
+use std::time::Duration;
+
 use candid::Principal;
-use catalyze_shared::{api_error::ApiError, paged_response::PagedResponse, CanisterResult};
+use catalyze_shared::{paged_response::PagedResponse, CanisterResult, CellStorage};
 use common::{
-    is_developer, is_proxy, spawn_shard, CellStorage, IndexConfig, IndexConfigBase,
-    IndexController, ShardsIndex,
+    is_developer, is_proxy, IndexConfig, IndexConfigBase, IndexController, Principals, ShardsIndex,
 };
-use ic_cdk::{init, query, trap, update};
+use ic_cdk::{init, post_upgrade, query, trap, update};
 use serde_bytes::ByteBuf;
 
 use crate::{
@@ -27,28 +28,31 @@ fn init(proxies: Vec<Principal>) {
         trap("Proxies cannot be empty");
     }
 
-    config()
-        .proxies()
-        .set(proxies.into())
-        .expect("Failed to set proxies");
+    set_proxies(proxies).expect("Failed to set proxies");
+}
+
+#[post_upgrade]
+fn post_upgrade() {
+    ic_cdk_timers::set_timer(Duration::from_secs(1), || {
+        ic_cdk::spawn(async {
+            controller().init_timers().await;
+        });
+    });
+}
+
+#[update(guard = "is_proxy_guard")]
+fn set_proxies(proxies: Vec<Principal>) -> CanisterResult<Principals> {
+    config().proxies().set(proxies.into())
+}
+
+#[query(guard = "is_proxy_guard")]
+fn _dev_get_shards() -> CanisterResult<ShardsIndex> {
+    config().shards().get()
 }
 
 #[update(guard = "is_proxy_guard")]
 async fn _dev_extend_shards(shards: u64) -> CanisterResult<ShardsIndex> {
-    let shard_ids = config().shards().get().unwrap_or_default();
-    let shard_wasm = config().shard_wasm().get()?;
-    let mut new_shards_list = shard_ids.to_vec();
-
-    for _ in 0..shards {
-        new_shards_list.push(spawn_shard(shard_wasm.clone()).await?);
-    }
-    let shard_ids = config().shards().set(new_shards_list.clone().into())?;
-
-    if config().shard_iter().get().is_err() {
-        config().shard_iter().set(new_shards_list[0].id())?;
-    }
-
-    Ok(shard_ids)
+    controller().extend_shards(shards).await
 }
 
 #[update(guard = "is_proxy_guard")]
@@ -58,19 +62,12 @@ fn _dev_upload_wasm(wasm: ByteBuf) -> bool {
 
 #[update(guard = "is_proxy_guard")]
 fn _dev_set_shard_filled(shard: Principal, filled: bool) -> CanisterResult<ShardsIndex> {
-    let mut shard_ids = config().shards().get()?.to_vec();
+    controller().set_shard_filled(shard, filled)
+}
 
-    let idx = shard_ids
-        .iter()
-        .position(|s| s.id() == shard)
-        .ok_or_else(|| {
-            ApiError::not_found().add_message(format!("Shard with the id {shard} not found"))
-        })?;
-
-    let shard = shard_ids.get_mut(idx).expect("Shard not found");
-    shard.set_filled(filled);
-
-    config().shards().set(shard_ids.clone().into())
+#[update(guard = "is_proxy_guard")]
+async fn _dev_upgrade_shard(shard: Principal) -> CanisterResult<()> {
+    controller().upgrade_shard(shard).await
 }
 
 #[query(composite = true, guard = "is_proxy_guard")]
@@ -130,8 +127,13 @@ async fn insert(key: Key, value: Value) -> CanisterResult<Entry> {
 }
 
 #[update(guard = "is_proxy_guard")]
+async fn insert_many(list: Vec<Entry>) -> CanisterResult<Vec<Entry>> {
+    controller().insert_many(list).await
+}
+
+#[update(guard = "is_proxy_guard")]
 async fn update(key: Key, value: Value) -> CanisterResult<Entry> {
-    controller().update(key, value).await
+    controller().update_profile(key, value).await
 }
 
 #[update(guard = "is_proxy_guard")]