+++ src/shard_group/src/aliases.rs
@@ -1,7 +1,7 @@
 use catalyze_shared::group_with_members::{GroupEntry, GroupFilter, GroupWithMembers};
 
 #[allow(dead_code)]
-pub const DATA_KIND: &str = "shard_group";
+pub const DATA_KIND: &str = "group";
 
 pub type Key = u64;
 pub type Value = GroupWithMembers;