+++ CHANGELOG.md
@@ -15,3 +15,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Report index and shard implementations
 - Pagination support
 - Boosted index implementation
+- Notification index and shard implementations
+- Topic index implementation